Major Refactor of Code
