+++ class/sub_distributors.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
-# Subdistributor class that extends distributor
+require_relative '../helpers/input_helper'
+require_relative '../helpers/data_helper'
+require_relative '../helpers/permission_helper'
+
+# Subdistributor class that extends distributor class
 class Subdistributors < Distributors
   attr_accessor :permissible_data, :main_distributor_permissible_list
 
@@ -14,30 +18,23 @@ class Subdistributors < Distributors
   # Returns permissible data for an given sub-distributor
   def permissible_data_hash_sub_distributor
     removed_hash = remove_excluded_data
-    updated_hash = add_included_data(removed_hash)
-  end
-
-  # Returns only the required country for sub distributor list
-  def add_included_data(hash_data)
-    updated_hash = {}
-    include_list.each do |included_region|
-      updated_hash.merge!(included_region['countries'] => hash_data[included_region['countries']])
-    end
-    updated_hash
+    update_included_list(removed_hash)
   end
 
   # Removes excluded data for an sub distributor
   def remove_excluded_data
     updated_hash_data = @main_distributor_permissible_list
+    return updated_hash_data if exclude_list.empty?
+
     exclude_list.each do |excluded_region|
       next if excluded_region['countries'].nil?
 
       if excluded_region['province'].nil?
         updated_hash_data.delete(excluded_region['countries'])
       elsif excluded_region['cities'].nil?
-        updated_hash_data[excluded_region['countries']]['province'].delete(excluded_region['province'])
+        updated_hash_data[excluded_region['countries']].delete(excluded_region['province'])
       else
-        updated_hash_data[excluded_region['countries']]['province'][excluded_region['province']]['cities']
+        updated_hash_data[excluded_region['countries']][excluded_region['province']]
           .delete(excluded_region['cities'])
       end
     end
@@ -50,14 +47,14 @@ class Subdistributors < Distributors
     if permissible_data[region_to_check['countries']].nil?
       display_message(region_to_check['countries'], false)
     elsif region_to_check['province'] &&
-          permissible_data[region_to_check['countries']]['province'][region_to_check['province']].nil?
+          permissible_data[region_to_check['countries']][region_to_check['province']].nil?
       display_message(region_to_check['province'], false)
     elsif region_to_check['cities'] &&
-          permissible_data[region_to_check['countries']]['province'][region_to_check['province']]['cities'][region_to_check['cities']]
+          permissible_data[region_to_check['countries']][region_to_check['province']][region_to_check['cities']]
           .nil?
       display_message(region_to_check['cities'], false)
     else
-      display_message(region_to_check['countries'], true)
+      display_message('', true)
     end
     nil
   end
